run-time: classification, range, partition: store size using int instead of size_t, to avoid signed/unsigned gotcha in comparisons in model code

+++ include/omc/Classification.h
@@ -108,7 +108,7 @@ public:
     // limits (static constants)
     static const T min = (T) 0;
     static const T max = (T) (T_size - 1);
-    static const size_t size = T_size;
+    static const int size = T_size;
 
 private:
     // cover function to get value
